+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Features:
   - Server : Configurable allowed origins for the websocket server in docker-compose file
 - Bugfixes:
+  - Unwanted space at the bottom or reference role modal
 - Localization :
 - Technical :